+++ .tool-versions
@@ -1,0 +1,2 @@
+erlang 22.0.5
+elixir 1.9.1
