--- lib/algae/free/monad.ex
@@ -1,4 +1,0 @@
-import TypeClass
-
-definst Witchcraft.Monad, for: Algae.Free.Pure
-definst Witchcraft.Monad, for: Algae.Free.Roll
